runtime: Refactor
