+++ graph/src/runtime/asc_ptr.rs
@@ -1,6 +1,6 @@
-use crate::error::DeterministicHostError;
+use super::DeterministicHostError;
 
-use super::{class::EnumPayload, AscHeap, AscType, AscValue};
+use super::{AscHeap, AscType};
 use std::fmt;
 use std::marker::PhantomData;
 use std::mem::size_of;
@@ -30,9 +30,10 @@ impl<T> fmt::Debug for AscPtr<T> {
 
 impl<C> AscPtr<C> {
     /// A raw pointer to be passed to Wasm.
-    pub(crate) fn wasm_ptr(self) -> u32 {
+    pub fn wasm_ptr(self) -> u32 {
         self.0
     }
+
     #[inline(always)]
     pub fn new(heap_ptr: u32) -> Self {
         Self(heap_ptr, PhantomData)
@@ -42,18 +43,18 @@ impl<C> AscPtr<C> {
 impl<C: AscType> AscPtr<C> {
     /// Create a pointer that is equivalent to AssemblyScript's `null`.
     #[inline(always)]
-    pub(crate) fn null() -> Self {
+    pub fn null() -> Self {
         AscPtr::new(0)
     }
 
     /// Read from `self` into the Rust struct `C`.
-    pub(super) fn read_ptr<H: AscHeap>(self, heap: &H) -> Result<C, DeterministicHostError> {
+    pub fn read_ptr<H: AscHeap>(self, heap: &H) -> Result<C, DeterministicHostError> {
         let bytes = heap.get(self.0, C::asc_size(self, heap)?)?;
         C::from_asc_bytes(&bytes)
     }
 
     /// Allocate `asc_obj` as an Asc object of class `C`.
-    pub(super) fn alloc_obj<H: AscHeap>(
+    pub fn alloc_obj<H: AscHeap>(
         asc_obj: &C,
         heap: &mut H,
     ) -> Result<AscPtr<C>, DeterministicHostError> {
@@ -62,7 +63,7 @@ impl<C: AscType> AscPtr<C> {
     }
 
     /// Helper used by arrays and strings to read their length.
-    pub(super) fn read_u32<H: AscHeap>(&self, heap: &H) -> Result<u32, DeterministicHostError> {
+    pub fn read_u32<H: AscHeap>(&self, heap: &H) -> Result<u32, DeterministicHostError> {
         // Read the bytes pointed to by `self` as the bytes of a `u32`.
         let raw_bytes = heap.get(self.0, size_of::<u32>() as u32)?;
         let mut u32_bytes: [u8; size_of::<u32>()] = [0; size_of::<u32>()];
@@ -71,17 +72,17 @@ impl<C: AscType> AscPtr<C> {
     }
 
     /// Conversion to `u64` for use with `AscEnum`.
-    pub(crate) fn to_payload(&self) -> u64 {
+    pub fn to_payload(&self) -> u64 {
         self.0 as u64
     }
 
     /// We typically assume `AscPtr` is never null, but for types such as `string | null` it can be.
-    pub(crate) fn is_null(&self) -> bool {
+    pub fn is_null(&self) -> bool {
         self.0 == 0
     }
 
     // Erase type information.
-    pub(crate) fn erase(self) -> AscPtr<()> {
+    pub fn erase(self) -> AscPtr<()> {
         AscPtr::new(self.0)
     }
 }
@@ -89,18 +90,6 @@ impl<C: AscType> AscPtr<C> {
 impl<C> From<u32> for AscPtr<C> {
     fn from(ptr: u32) -> Self {
         AscPtr::new(ptr)
-    }
-}
-
-impl<C> From<EnumPayload> for AscPtr<C> {
-    fn from(payload: EnumPayload) -> Self {
-        AscPtr::new(payload.0 as u32)
-    }
-}
-
-impl<C> From<AscPtr<C>> for EnumPayload {
-    fn from(x: AscPtr<C>) -> EnumPayload {
-        EnumPayload(x.0 as u64)
     }
 }
 
@@ -114,5 +103,3 @@ impl<T> AscType for AscPtr<T> {
         Ok(AscPtr::new(bytes))
     }
 }
-
-impl<T> AscValue for AscPtr<T> {}
