+++ runtime/wasm/src/to_from/mod.rs
@@ -2,9 +2,11 @@ use std::collections::HashMap;
 use std::hash::Hash;
 use std::iter::FromIterator;
 
+use graph::runtime::{
+    AscHeap, AscPtr, AscType, AscValue, DeterministicHostError, FromAscObj, ToAscObj, TryFromAscObj,
+};
+
 use crate::asc_abi::class::*;
-use crate::asc_abi::{AscHeap, AscPtr, AscType, AscValue, FromAscObj, ToAscObj, TryFromAscObj};
-use crate::error::DeterministicHostError;
 
 ///! Implementations of `ToAscObj` and `FromAscObj` for Rust types.
 ///! Standard Rust types go in `mod.rs` and external types in `external.rs`.
